Reorganized files into corresponding folders
