+++ JavaGUI/firstscene.css
@@ -1,0 +1,34 @@
+#root{
+    -fx-background-image: url("https://img.freepik.com/free-vector/cerulean-blue-curve-frame-template-vector_53876-136094.jpg?w=740&t=st=1669345625~exp=1669346225~hmac=5215b932f8a81ca6f9035f38bc1a2382a45ef793d862e305b8e8b6e1a6522e7e");
+
+    -fx-background-size: cover;
+}
+
+#Welcomelabel {
+    -fx-font-size:30;
+    -fx-font-family: "Open Sans";
+    -fx-font-weight:bold;
+    -fx-text-fill:#2E5A88;
+    
+}
+#welcomeheart{
+    -fx-font-size:30;
+    -fx-font-weight:bold;
+    -fx-font-family: "Open Sans";
+    -fx-text-fill:#2E5A88;
+}
+
+#signin {
+    -fx-font-size:13;
+    -fx-font-weight:bold;
+    -fx-font-family: "Open Sans";
+    -fx-background-radius: 30;
+    -fx-border-radius: 30;
+}
+#signup {
+    -fx-font-size:13;
+    -fx-font-weight:bold;
+    -fx-font-family: "Open Sans";
+    -fx-background-radius: 30;
+    -fx-border-radius: 30;
+}
