--- PicoProject/Mainscene.fxml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-    Copyright (c) 2017, Gluon and/or its affiliates.
-    All rights reserved. Use is subject to license terms.
-    This file is available and licensed under the following license:
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions
-    are met:
-   - Redistributions of source code must retain the above copyright
-     notice, this list of conditions and the following disclaimer.
-   - Redistributions in binary form must reproduce the above copyright
-     notice, this list of conditions and the following disclaimer in
-     the documentation and/or other materials provided with the distribution.
-   - Neither the name of Oracle Corporation and Gluon nor the names of its
-     contributors may be used to endorse or promote products derived
-     from this software without specific prior written permission.
-  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
--->
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ButtonBar?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane fx:id="root" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="506.0" prefWidth="275.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="hellofx.MainSceneController">
-   <children>
-      <Label fx:id="Welcomelabel" layoutX="14.0" layoutY="22.0" prefHeight="33.0" prefWidth="219.0" text="Welcome to" />
-      <Label fx:id="welcomeheart" layoutX="31.0" layoutY="55.0" prefHeight="24.0" prefWidth="239.0" text="\@heartrate" />
-      <Button fx:id="signin" layoutX="18.0" layoutY="326.0" mnemonicParsing="false" onAction="#siginscene" prefHeight="30.0" prefWidth="239.0" text="signin" />
-      <Button fx:id="signup" layoutX="18.0" layoutY="365.0" mnemonicParsing="false" onAction="#opensignuplink" prefHeight="33.0" prefWidth="239.0" text="signup" />
-      <ButtonBar layoutX="-12.0" layoutY="398.0" prefHeight="40.0" prefWidth="200.0" />
-   </children>
-</AnchorPane>
