--- PicoProject/atsignsignin.fxml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-    Copyright (c) 2017, Gluon and/or its affiliates.
-    All rights reserved. Use is subject to license terms.
-    This file is available and licensed under the following license:
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions
-    are met:
-   - Redistributions of source code must retain the above copyright
-     notice, this list of conditions and the following disclaimer.
-   - Redistributions in binary form must reproduce the above copyright
-     notice, this list of conditions and the following disclaimer in
-     the documentation and/or other materials provided with the distribution.
-   - Neither the name of Oracle Corporation and Gluon nor the names of its
-     contributors may be used to endorse or promote products derived
-     from this software without specific prior written permission.
-  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
--->
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane fx:id="scene2" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="506.0" prefWidth="275.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="hellofx.atsignsigninController">
-   <children>
-      <Button fx:id="backmainscene" layoutX="7.0" layoutY="2.0" mnemonicParsing="false" onAction="#gobacktosignin" prefHeight="28.0" prefWidth="36.0" text="←">
-         <font>
-            <Font name="System Bold" size="16.0" />
-         </font></Button>
-      <Label fx:id="loginnowlabel" layoutX="17.0" layoutY="43.0" prefHeight="63.0" prefWidth="241.0" text="Login Now">
-         <font>
-            <Font name="Microsoft Sans Serif" size="12.0" />
-         </font>
-      </Label>
-      <TextField fx:id="useratsigntextview" layoutX="17.0" layoutY="188.0" prefHeight="35.0" prefWidth="241.0" />
-      <TextField fx:id="machineatsigntextview" layoutX="17.0" layoutY="261.0" prefHeight="35.0" prefWidth="241.0" />
-      <Label fx:id="useratsignlabel" layoutX="25.0" layoutY="162.0" text="Your atsign" />
-      <Label fx:id="machineatsignlabel" layoutX="17.0" layoutY="236.0" text="heartrate machine's atsign">
-         <font>
-            <Font name="Microsoft Sans Serif" size="12.0" />
-         </font>
-      </Label>
-      <Button fx:id="loginconfirm" layoutX="199.0" layoutY="311.0" mnemonicParsing="false" onAction="#heartratedashboard" text="Login -&gt;" />
-   </children>
-</AnchorPane>
