--- PicoProject/secondscene.css
@@ -1,27 +1,0 @@
-#scene2 {
-    -fx-background-image: url("https://img.freepik.com/free-vector/cerulean-blue-curve-frame-template-vector_53876-136094.jpg?w=740&t=st=1669345625~exp=1669346225~hmac=5215b932f8a81ca6f9035f38bc1a2382a45ef793d862e305b8e8b6e1a6522e7e");
-
-    -fx-background-size: cover;
-}
-#backmainscene {
-    -fx-background-radius: 30;
-    -fx-border-radius: 30;
-}
-
-#loginnowlabel {
-    -fx-font-size:30;
-    -fx-font-weight:bold;
-    -fx-text-fill:#2E5A88;
-}
-#useratsigntextview{
-    -fx-background-radius: 30;
-    -fx-border-radius: 30;
-}
-#machineatsigntextview{
-    -fx-background-radius: 30;
-    -fx-border-radius: 30;
-}
-#loginconfirm {
-    -fx-background-radius: 30;
-    -fx-border-radius: 30;
-}
