--- PicoProject/thirdscene.css
@@ -1,26 +1,0 @@
-#heartrateroot{
-    -fx-background-image: url("https://img.freepik.com/free-vector/cerulean-blue-curve-frame-template-vector_53876-136094.jpg?w=740&t=st=1669345625~exp=1669346225~hmac=5215b932f8a81ca6f9035f38bc1a2382a45ef793d862e305b8e8b6e1a6522e7e");
-    -fx-background-size: cover;
-}
-
-#backscene3{
-    -fx-background-radius: 30;
-    -fx-border-radius: 30;
-    -fx-background-color:#00bfff;
-    -fx-text-fill:white;
-
-}
-
-#startmeasure{
-    -fx-background-color:#00bfff;
-    -fx-text-fill:white;
-}
-#saveval {
-    -fx-background-color:#00bfff;
-    -fx-text-fill:white;
-}
-#heartrateval{
-    -fx-font-size:50;
-    -fx-font-weight:bold;
-    -fx-text-fill:#00bfff;
-}
